myblog/requirements.txt
Django==2.0.8
django-crispy-forms==1.8.1
django-js-asset==1.2.2
django-mptt==0.10.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
